# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "VK TPA Website API Documentation",
        "title": "TPA-Website",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/flight/airline": {
            "get": {
                "description": "Get an airline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airline"
                ],
                "summary": "Get airline",
                "parameters": [
                    {
                        "description": "Airline",
                        "name": "airline",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Airline found successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new airline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airline"
                ],
                "summary": "Add airline",
                "parameters": [
                    {
                        "description": "Airline",
                        "name": "airline",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Airline created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/flight/schedule": {
            "get": {
                "description": "Get a list of flight schedules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight"
                ],
                "summary": "List flight schedules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.FlightSchedule"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new flight schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight"
                ],
                "summary": "Add flight schedule",
                "parameters": [
                    {
                        "description": "Flight Schedule",
                        "name": "flightSchedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight schedule created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flight/schedule/": {
            "put": {
                "description": "Update flight schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight"
                ],
                "summary": "Update flight schedule",
                "parameters": [
                    {
                        "description": "Flight Schedule",
                        "name": "flightSchedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight schedule updated successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/flight/schedule/{flightID}": {
            "get": {
                "description": "Get a flight schedule by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight"
                ],
                "summary": "Get flight schedule by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flight ID",
                        "name": "flightID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight schedule found successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flight/search/{query}": {
            "get": {
                "description": "Search flight schedules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flight"
                ],
                "summary": "Search flight schedules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query",
                        "name": "query",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Airline": {
            "type": "object",
            "properties": {
                "airlineID": {
                    "type": "integer"
                },
                "airlineLogo": {
                    "type": "string"
                },
                "airlineName": {
                    "type": "string"
                }
            }
        },
        "model.Airport": {
            "type": "object",
            "properties": {
                "airportCode": {
                    "type": "string"
                },
                "airportID": {
                    "type": "integer"
                },
                "airportLocation": {
                    "type": "string"
                },
                "airportName": {
                    "type": "string"
                }
            }
        },
...
        "model.FlightSchedule": {
            "type": "object",
            "properties": {
                "arrivalTime": {
                    "type": "string"
                },
                "businessPrice": {
                    "type": "integer"
                },
                "departureTime": {
                    "type": "string"
                },
                "destinationAirport": {
                    "$ref": "#/definitions/model.Airport"
                },
                "destinationAirportID": {
                    "type": "integer"
                },
                "economyPrice": {
                    "type": "integer"
                },
                "flightScheduleCode": {
                    "type": "string"
                },
                "flightScheduleID": {
                    "type": "integer"
                },
                "originAirport": {
                    "$ref": "#/definitions/model.Airport"
                },
                "originAirportID": {
                    "type": "integer"
                },
                "plane": {
                    "$ref": "#/definitions/model.Plane"
                },
                "planeID": {
                    "type": "integer"
                },
                "seats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Seat"
                    }
                }
            }
        },
...
        "model.Plane": {
            "type": "object",
            "properties": {
                "airline": {
                    "$ref": "#/definitions/model.Airline"
                },
                "airlineID": {
                    "type": "integer"
                },
                "businessSeats": {
                    "type": "integer"
                },
                "economySeats": {
                    "type": "integer"
                },
                "planeCode": {
                    "type": "string"
                },
                "planeID": {
                    "type": "integer"
                }
            }
        },
...
        "model.Seat": {
            "type": "object",
            "properties": {
                "flightScheduleID": {
                    "type": "integer"
                },
                "planeID": {
                    "type": "integer"
                },
                "seatID": {
                    "type": "integer"
                },
                "seatNumber": {
                    "type": "string"
                },
                "seatStatus": {
                    "type": "string"
                },
                "seatType": {
                    "type": "string"
                }
            }
        },
...
    },
...
}